requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.20

# Security
bcrypt==4.0.1
cryptography==41.0.4

# Validation
marshmallow==3.20.1
email-validator==2.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.1
redis==4.6.0

# Development
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Production
gunicorn==21.2.0
